# src/main/resources/static/swagger.yml
openapi: 3.0.4
info:
  title: itPlace Admin API Document
  description: |-
    지도 기반 내 주변 LG유플러스 멤버십 혜택 정보 조회 서비스 'itPlace'의 관리자 페이지 API 명세서입니다.
    
    모든 API 응답은 표준 응답 포맷(ApiResponse)을 따릅니다.
  version: 1.0.0
servers:
  - url: http://localhost:8081/api/v1
    description: 개발 서버
...
paths:
...
  /benefits/favorite:
    get:
      tags:
        - Dashboard
      summary: 전체 관심 혜택 통계
      description: 관심 혜택으로 많이 한 혜택들을 조회합니다.
      operationId: getFavoritesStatistics
      parameters:
        - name: limit
          in: query
          description: 조회할 혜택 개수
          required: false
          schema:
            type: integer
            default: 5
      responses:
        '200':
          description: 관심 혜택 통계 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                등급별 혜택 이용률 조회 성공:
                  value:
                    code: "BENEFITS_FAVORITES_SUCCESS"
                    status: "OK"
                    message: "즐겨찾기 통계가 성공적으로 조회되었습니다."
                    data:
                      - benefitId: 1
                        partnerName: "CGV"
                        benefitName: "2천원 할인"
                        favoriteCount: 156
                        rank: 1
                      - benefitId: 2
                        partnerName: "GS25"
                        benefitName: "1천원 당 100원 할인"
                        favoriteCount: 142
                        rank: 2
                      - benefitId: 3
                        partnerName: "배스킨라빈스"
                        benefitName: "패밀리 사이즈(5가지 맛) 4천원 할인"
                        favoriteCount: 128
                        rank: 3
                      - benefitId: 4
                        partnerName: "CGV"
                        benefitName: "영화 1+1"
                        favoriteCount: 95
                        rank: 4
                    timestamp: "2025-07-04T11:27:06.789Z"
...
  /benefit/{benefitId}:
    get:
      tags:
        - Benefit
      summary: 혜택 상세 조회
      description: 혜택 ID를 통해 혜택의 상세 정보를 조회합니다.
      operationId: getBenefitDetail
      parameters:
        - name: benefitId
          in: path
          required: true
          description: 조회할 혜택의 ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 혜택 상세 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BenefitDetailResponse'
              example:
                data:
                  benefitId: 148
                  benefitName: "아르떼 키즈파크 제주"
                  description: "전 세계 육백만 명 이상이 관람한 아르떼뮤지엄이 선보이는 1,200평 규모의 체험형 미디어아트 놀이터"
                  benefitLimit: "제한없음"
                  manual: "■  쿠폰 사용 방법\n- 결제 시 현장에서 LG U+ 멤버십 카드 제시\n\n■ 혜택 적용점 및 혜택 기준\n- 아르떼 키즈파크 제주점 \n- 이용권 20% 할인(동반 3인까지)\n\n■ 유의 사항\n- 선착순 쿠폰 소진, 전산 에러 등으로 인해 쿠폰 발급 불가 시 대체 혜택을 받으실 수 없습니다.\n- 당사 또는 제휴사 사정으로 인해 혜택이 조기 종료될 수 있습니다.\n- 해당 쿠폰은 제휴 기간인 2025.07.01 ~ 2025.12.31까지 사용 가능합니다.\n- 본 쿠폰은 아르떼 키즈파크 제주에서 본인 포함 최대 4인까지 사용 가능합니다.\n- 제주도민, 국가유공자 등 할인대상자 중복 사용 불가합니다\n\n■ 대표 문의\nTEL: 제주 064-799-7007 \nHOME: https://www.artekidspark.com"
                  url: "https://www.artekidspark.com/"
                  partnerName: "아르떼 키즈파크 제주"
                  image: "https://itplacepartners.s3.ap-northeast-2.amazonaws.com/img/arte.png"
                  tierBenefits:
                    - grade: "BASIC"
                      context: "이용권 최대 20% 할인(본인 포함 최대 4인까지)"
                      isAll: true
                    - grade: "VIP"
                      context: "이용권 최대 20% 할인(본인 포함 최대 4인까지)"
                      isAll: true
                    - grade: "VVIP"
                      context: "이용권 최대 20% 할인(본인 포함 최대 4인까지)"
                      isAll: true
                code: BENEFIT_DETAIL_SUCCESS
                status: OK
                message: 혜택 상세 조회에 성공했습니다.
                timestamp: '2025-07-16T12:00:00.000Z'

        '404':
          description: 혜택을 찾을 수 없습니다
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                data: null
                code: BENEFIT_NOT_FOUND
                status: NOT_FOUND
                message: 존재하지 않는 혜택입니다.
                timestamp: '2025-07-16T12:00:00.000Z'

    put:
      tags:
        - Benefit
      summary: 혜택 상세 정보 수정
      description: 특정 혜택의 정보를 수정합니다.
      operationId: updateBenefit
      parameters:
        - name: benefitId
          in: path
          description: 수정할 혜택 ID
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BenefitUpdateRequest'
            example:
              benefitLimit: "월 2회"
              manual: |
                기술의 발전은 우리가 세상을 이해하고 상호작용하는 방식을 근본적으로 변화시켰으며,\n이는 단순한 편리함을 넘어 사회 구조와 개인의 삶에 깊은 영향을 미치고 있습니다.\n특히 인공지능과 데이터 과학 분야의 비약적인 성장은 미래의 일자리 시장은 물론,\n윤리적 고려와 개인 정보 보호의 중요성을 끊임없이 제기하며\n새로운 논의의 장을 열고 있습니다.\n1 > 2 > 3 수정합니다.
      responses:
        '200':
          description: 혜택 정보 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                data: null
                code: BENEFIT_UPDATE_SUCCESS
                status: OK
                message: 혜택 정보가 성공적으로 수정되었습니다.
                timestamp: "2025-07-23T12:00:00.000Z"
        '400':
          description: 수정할 내용이 없는 경우
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                data: null
                code: BENEFIT_NOT_UPDATED
                status: BAD_REQUEST
                message: 변경된 값이 없습니다.
                timestamp: "2025-07-23T12:00:00.000Z"
        '404':
          description: 존재하지 않는 혜택 ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                data: null
                code: BENEFIT_NOT_FOUND
                status: NOT_FOUND
                message: 존재하지 않는 혜택입니다.
                timestamp: "2025-07-23T12:00:00.000Z"
  /benefits:
    get:
      tags:
        - Benefit
      summary: 전체 제휴처 목록 조회 -- 구상 중
      description: 전체 제휴처 목록을 조회합니다.
      operationId: getAllBenefits
      parameters:
        - name: mainCategory
          in: query
          required: false
          description: 카테고리 ("기본혜택, VIP콕")
          schema:
            type: string
            enum: ["BASIC_BENEFIT", "VIP_COCK"]
        - name: category
          in: query
          required: false
          description: 카테고리 (푸드, 엑티비티 등)
          schema:
            type: string
        - name: type
          in: query
          required: false
          description: 혜택유형(증정, 할인)
          schema:
            type: string
            enum: ["FREE", "DISCOUNT"]
        - name: keyword
          in: query
          required: false
          description: 검색어
          schema:
            type: string
        - name: page
          in: query
          required: false
          description: 페이지 번호 (0부터 시작)
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          description: 페이지 크기
          schema:
            type: integer
            format: int32
            default: 8
        - name: sortBy
          in: query
          required: false
          description: 정렬 기준("필드")
          schema:
            type: string
            default: id
        - name: direction
          in: query
          required: false
          description: 정렬 기준("방향")
          schema:
            type: string
            default: asc
      responses:
        '200':
          description: 제휴처 목록 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                제휴처 목록 조회 성공:
                  value:
                    code: "BENEFITS_LIST_SUCCESS"
                    status: "OK"
                    message: "제휴처 목록이 성공적으로 조회되었습니다."
                    data:
                      content:
                        - benefitId: 1
                          benefitName: "CGV"
                          mainCategory: "VIP_COCK"
                          category: "문화/여가"
                          type: "FREE"
                          image: "https://example.com/image.png"
                          searchRank: 4
                          favoriteRank: 2
                          usageRank: 6
                        - benefitId: 1
                          benefitName: "CGV"
                          mainCategory: "BASIC_BENEFIT"
                          category: "문화/여가"
                          type: "DISCOUNT"
                          image: "https://example.com/image.png"
                          searchRank: 4
                          favoriteRank: 1
                          usageRank: 2
                      page: 0
                      totalPages: 5
                      totalElements: 40
                      hasNext: true
                      code: BENEFIT_PAGE_SUCCESS
                      status: OK
                      message: 제휴처 목록 조회에 성공했습니다.
                      timestamp: '2025-07-17T12:00:00.000Z'
...
components:
  schemas:
...
    ApiResponse:
      type: object
      properties:
        code:
          type: string
          description: "서버에서 정의한 응답 코드"
          example: "SUCCESS_200_1"
        status:
          type: string
          description: "HTTP 상태 메시지"
          example: "OK"
        message:
          type: string
          description: "응답에 대한 설명 메시지"
          example: "요청이 정상적으로 처리되었습니다."
        data:
          type: object
          description: "실제 응답 데이터 (nullable)"
          nullable: true
        timestamp:
          type: string
          format: date-time
          description: "응답 시간"

    # ----------------------------------------------------
    # Request DTOs
    # ----------------------------------------------------
...
    BenefitUpdateRequest:
      type: object
      properties:
        benefitLimit:
          type: string
        manual:
          type: string

    # ----------------------------------------------------
    # Response DTOs
    # ----------------------------------------------------
...
    BenefitDetailResponse:
      type: object
      properties:
        benefitId:
          type: integer
          format: int64
        benefitName:
          type: string
        description:
          type: string
        benefitLimit:
          type: string
        manual:
          type: string
        url:
          type: string
          format: uri
        partnerName:
          type: string
        image:
          type: string
          format: uri
        tierBenefits:
          type: array
          items:
            $ref: '#/components/schemas/TierBenefitInfo'
    TierBenefitInfo:
      type: object
      properties:
        grade:
          type: string
          enum: [BASIC, VIP, VVIP, VIP콕]
        context:
          type: string
        isAll:
          type: boolean
...